refactor : split OutFormat && Labels for print
